Throttled SystemSettings saving
Every time a system setting is saved a call to saveChanges is made.
This becomes wasteful/problematic when autoSave is checked and there are
several persistent options and one or several of them are changed within
a short period of time, e.g. clicking lots of checkboxes in the tweak menu
or a plugin setting lots of config options that are named in SystemSettings

On the client with autoSave on a big TiddlyWiki this could lead to browser
freezing due to the expensive save operation being run too frequently.

On a server such as TiddlySpace this can lead to edit conflicts.

This proposed change simply adds a commitSystemSettings function which
uses a timeout to manage frequent calls to itself.