do not dirty the story when invoking loadMissingTiddler on none file uri
On a server load missing tiddler retrieves content that is already saved somewhere
thus it doesnt need to make the store dirty and can cause problems

(see https://github.com/TiddlySpace/tiddlyspace/issues/630)

This suggested solution checks the store state before saving in the local store and reverts to it when
not on a file uri.

Would benefit from review of someone who understands the adaptor mechanism inside out.