updated timeline macro to support templating
this extends the timeline macro to accept the following parameters
* filter
* template
* groupTemplate

this allows the timeline to work in a similar way to the list macro
which also supports templating of this style.

template is the name of a tiddler template that will be wikified in the context
of the current timeline list item.