Fix issues #20: httpSuccess no longer present
This is temporary fix to the fact that jquery 1.6.x does not
include jQuery.httpSuccess() method. The fix is to copy in
the code from old jQuery. This is only a temporary fix as
the enclosing httpReq method is deprecated anyway, so if the fix
in this trac ticket is taken to completion then the problem
goes away:

   http://trac.tiddlywiki.org/ticket/1268