Move config.tasks code into backstage
The backstage introduces and makes use of the config.tasks codebase.
functionality is tied to it in Lingo.js (the saveChanges action is assigned to it)

All this would be better served by having this code in Backstage.js, as code which does not
use the backstage has no need for it.