extend test coverage
This creates failing tests and reveal a flaw in the importTiddlyWiki mechanism

In TiddlySpace where ajax is not possible due to cross domain restrictions
content is rendered into an iframe, the html of the iframe is obtained and
it is passed through the importTiddlyWiki function

When rendering content in iframes, the html rendered does not always
match the original HTML. Browsers (IE6 for instance) may upper case tag names ie. div becomes
DIV meaning that string searching cannot be relied on outside of ajax
requests.

TiddlySpace has encountered this problem:
see https://github.com/TiddlySpace/tiddlyspace/issues/675

Ideally we should be doing dom searching rather than string searching - the only thing
that should be required is an element with id storeArea