describe behaviour timeline 1st parameter with prefix
the 1st parameter of the timeline points to a field that should be
used to construct the timeline. If given a prefix of - it should
reverse the order of the timeline. Currently this is broken
see http://groups.google.com/group/tiddlywikidev/browse_thread/thread/8425d68ee54c8ade

this test case covers this scenario and describes the expected behaviour
the update to js/Macro.js makes the test pass