added consist. about white-space around slicenames
Now [in table notation] all white-space around slice-names is stripped off (the first and the last character can't be a white-space symbol \s).

One leading space is allowed before all the wrappers ("//", "''" and "~")
and arbitrary amount of trailing tabs/spaces (after "//", "''" or ":") is as well -
so in the view mode one can have any text-alignment inside cells
and in the edit mode one can make the slices "table" look like a table by aligning "|" with spaces and tabs.